Simplifying code, making parameter non-constant and couple more tests for nextafter
