+++ src/function/scalar/math/CMakeLists.txt
@@ -1,5 +1,4 @@
-add_library_unity(duckdb_func_math OBJECT numeric.cpp random.cpp nextafter.cpp
-                  setseed.cpp)
+add_library_unity(duckdb_func_math OBJECT numeric.cpp random.cpp setseed.cpp)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_func_math>
     PARENT_SCOPE)
